add sbe9 file reader
This reader can read the sample file from Bob Key, some additional
features would be nice in the future, but require some more sample files
to be able to read.

Unresolved is how to handle duplicate parameters such as salinity and
temperature. There is also no bottom depth in the file.