plot etopo: change maridian/parallel drawer to use tuples and None for 0s
The "dash list" is a list of "on/off" pairs for how to draw a line on a
plot (in our case, a map). It seems it was discovered that certain
backends (agg) would infinite loop with any of the parameters being 0.
The docs state that a tuple of (None, None) results in a solid line.
It was discovered in testing that the list [None, None] did not work. So
we are switching to tuples here, guessing this has to do with tuples
being immutable.